feat: split trust-protection into protection and recovery.
logion-network/logion-internal#1197
