+++ src/legal-officer/ProtectedUsers.test.tsx
@@ -4,7 +4,7 @@ jest.mock('./LegalOfficerContext');
 import ProtectedUsers from "./ProtectedUsers";
 import { shallowRender } from '../tests';
 import { setActivatedProtectionRequests } from "./__mocks__/LegalOfficerContextMock";
-import { ACTIVATED_PROTECTION_REQUEST1, ACTIVATED_PROTECTION_REQUEST3 } from "../wallet-user/trust-protection/TestData";
+import { ACTIVATED_PROTECTION_REQUEST1, ACTIVATED_PROTECTION_REQUEST3 } from "../wallet-user/protection/TestData";
 
 test('renders one row', () => {
     setActivatedProtectionRequests([ ACTIVATED_PROTECTION_REQUEST1 ]);
