+++ src/wallet-user/UserRouter.tsx
@@ -34,8 +34,8 @@ import DashboardCertificateRouter from "../loc/DashboardCertificateRouter";
 import { useCommonContext } from "../common/CommonContext";
 import { useUserContext } from "./UserContext";
 import Home from "./home/Home";
-import TrustProtection from "./trust-protection/TrustProtection";
-import Recovery from "./trust-protection/Recovery";
+import TrustProtection from "./protection/TrustProtection";
+import Recovery from "./recovery/Recovery";
 import { useLogionChain } from '../logion-chain';
 import IdentityLocRequest from "../loc/IdentityLocRequest";
 import IssuerDashboard from "./issuer/IssuerDashboard";
