+++ src/wallet-user/protection/TrustProtection.tsx
