+++ src/wallet-user/recovery/VaultRecoveryProcessTab.css
