--- src/wallet-user/trust-protection/__mocks__/Model.tsx
@@ -1,3 +1,0 @@
-export function checkActivation() {
-    return Promise.resolve();
-}
